# specs/001-nutrition-recipe-planner/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Cookfully API
  version: 0.2.0
  description: >-
    Canonical HTTP contract for recipes, honest nutrition estimates, goals, meal plans,
    grocery lists, exports, agent access, and later goal-aware suggestions. Public decimal
    values are canonical strings. Estimated nutrition and suggestion projections are planning
    aids, not medical advice. MCP tools compose the same application commands and queries.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
  - bearerToken: []
...
paths:
...
  /nutrition-intelligence/settings:
    get:
      tags: [Nutrition Intelligence]
      operationId: getNutritionIntelligenceSettings
      responses:
        "200":
          description: Installation-level nutrition matching settings.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NutritionIntelligenceSettings" }
    put:
      tags: [Nutrition Intelligence]
      operationId: updateNutritionIntelligenceSettings
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NutritionIntelligenceSettingsWrite" }
      responses:
        "200":
          description: Updated nutrition matching settings.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NutritionIntelligenceSettings" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
...
  /nutrition-intelligence/estimate:
    post:
      tags: [Nutrition Intelligence]
      operationId: estimateNutritionIntelligence
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NutritionIntelligenceEstimateRequest" }
      responses:
        "200":
          description: Conservative model and concurrency resource estimate.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NutritionIntelligenceEstimate" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: cookfully_session
    bearerToken:
      type: http
      scheme: bearer
...
  responses:
...
    Conflict:
      description: Stale version, duplicate idempotency key with different input, or invalid transition.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    ValidationError:
      description: One or more fields are invalid.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    NutritionIntelligenceSettings:
      type: object
      additionalProperties: false
      required: [backend, intelligenceEnabled, inlineEnabled, modelName, modelRevision, concurrency, version, runtimeStatus]
      properties:
        backend: { type: string, enum: [hashing, fastembed] }
        intelligenceEnabled: { type: boolean, default: true }
        inlineEnabled: { type: boolean, default: true }
        modelName: { type: string, minLength: 1, maxLength: 200 }
        modelRevision: { type: [string, "null"] }
        concurrency: { type: integer, minimum: 1, maximum: 4 }
        version: { type: integer, minimum: 1 }
        runtimeStatus: { type: string, enum: [ready, configured, downloading, failed] }
        downloadJobId: { type: [string, "null"], format: uuid }
        downloadJobStatus: { type: [string, "null"] }
        downloadProgressCurrent: { type: [integer, "null"], minimum: 0 }
        downloadProgressTotal: { type: [integer, "null"], minimum: 0 }
        downloadFailureMessage: { type: [string, "null"] }
    NutritionIntelligenceSettingsWrite:
      type: object
      additionalProperties: false
      required: [backend, modelName, concurrency, version, estimateHash]
      properties:
        backend: { type: string, enum: [hashing, fastembed] }
        intelligenceEnabled: { type: boolean }
        inlineEnabled: { type: boolean }
        modelName: { type: string, minLength: 1, maxLength: 200 }
        concurrency: { type: integer, minimum: 1, maximum: 4 }
        version: { type: integer, minimum: 1 }
        estimateHash: { type: string, minLength: 64, maxLength: 64 }
    NutritionIntelligenceEstimateRequest:
      type: object
      additionalProperties: false
      required: [backend, modelName, concurrency]
      properties:
        backend: { type: string, enum: [hashing, fastembed] }
        modelName: { type: string, minLength: 1, maxLength: 200 }
        concurrency: { type: integer, minimum: 1, maximum: 4 }
...
    NutritionIntelligenceEstimate:
      type: object
      additionalProperties: false
      required: [backend, modelName, modelRevision, concurrency, activeFoodCount, downloadBytes, diskBytes, modelMemoryBytes, perJobMemoryBytes, totalMemoryBytes, requiredCpuCores, availableCpuCores, availableMemoryBytes, availableDiskBytes, memoryHeadroomBytes, status, warnings, estimateHash]
      properties:
        backend: { type: string, enum: [hashing, fastembed] }
        modelName: { type: string }
        modelRevision: { type: [string, "null"] }
        concurrency: { type: integer }
        activeFoodCount: { type: integer }
        downloadBytes: { type: integer }
        diskBytes: { type: integer }
        modelMemoryBytes: { type: integer }
        perJobMemoryBytes: { type: integer }
        totalMemoryBytes: { type: integer }
        requiredCpuCores: { type: integer }
        availableCpuCores: { type: integer }
        availableMemoryBytes: { type: integer }
        availableDiskBytes: { type: integer }
        memoryHeadroomBytes: { type: integer }
        status: { type: string, enum: [safe, warning, blocked] }
        warnings: { type: array, items: { type: string } }
        estimateHash: { type: string, minLength: 64, maxLength: 64 }
...
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        fieldErrors:
          type: array
          items:
            type: object
            required: [field, code]
            properties:
              field: { type: string }
              code: { type: string }
              message: { type: string }
...